requirements.txt
PySide6>=6.5.0
cryptography>=41.0.0
qrcode>=7.4.0
pyotp>=2.9.0
Pillow>=10.0.0
reportlab>=4.0.0
bcrypt>=4.1.0
keyring>=24.0.0
requests>=2.31.0
psutil>=5.9.0
# pyautogui removed - export system will be rebuilt
